+++ src/oliview/string_slice.h
@@ -8,22 +8,22 @@ namespace oliview {
         StringSlice();
         StringSlice(const Ptr<const std::string> & base);
         StringSlice(const Ptr<const std::string> & base,
-                    int offset,
-                    int length);
+                    size_t offset,
+                    size_t length);
         
         RHETORIC_GETTER(Ptr<const std::string>, base)
-        RHETORIC_GETTER(int, offset)
+        RHETORIC_GETTER(size_t, offset)
         
         const char * c_str() const;        
-        RHETORIC_GETTER(int, length)
+        RHETORIC_GETTER(size_t, length)
         
-        StringSlice Slice(int offset) const;
-        StringSlice Slice(int offset, int length) const;
+        StringSlice Slice(size_t offset) const;
+        StringSlice Slice(size_t offset, size_t length) const;
         
         std::string AsString() const;
     private:
         Ptr<const std::string> base_;
-        int offset_;
-        int length_;
+        size_t offset_;
+        size_t length_;
     };
 }
